+++ styleguide.config.js
@@ -8,14 +8,14 @@ const dir = path.join(__dirname, 'src');
 
 const portedComponents = [
   'badge', 'labeled-value', 'icon', 'logo', 'flag', 'avatar',
-  'radio-group', 'input', 'legend',
+  'radio-group', 'input', 'legend', 'dropdown',
 ];
 
 const quickFixedComponents = [
   'table', 'tbody', 'thead', 'tfoot', 'tr', 'th', 'td',
   'button', 'graph-tooltip', 'horizontal-nav', 'pane',
-  'dropdown', 'alert', 'nav-bar', 'tooltip', 'search',
-  'spinner', 'widget', 'range-selector', 'ratio-bar', 'select',
+  'alert', 'nav-bar', 'tooltip', 'search', 'select',
+  'spinner', 'widget', 'range-selector', 'ratio-bar',
 ];
 
 const allComponents = [...portedComponents, ...quickFixedComponents];