+++ CHANGELOG.md
@@ -11,13 +11,14 @@
 * Removed variables
 
 ### Ported components:
-Badge, LabeledValue, Icon, Logo, Flag, RadioGroup, Input
+Badge, LabeledValue, Icon, Logo, Flag, RadioGroup, Input,
+Legend
 
 ### Quickfix ported components
 These components have removed the SCSS styling, but has not added it as JSS.
 
 Table, Tbody, Thead, Tfoot, Tr, Th, Td,
-Button, GraphTooltip, HorizontalNav, Legend, Pane,
+Button, GraphTooltip, HorizontalNav, Pane,
 Dropdown, Alert, NavBar, Tooltip, Search,
 Spinner, Widget, RangeSelector, RatioBar, Select