+++ CHANGELOG.md
@@ -12,13 +12,13 @@
 
 ### Ported components:
 Badge, LabeledValue, Icon, Logo, Flag, SparkGraph,
-RadioGroup, Input
+RadioGroup, Input, Button,
 
 ### Quickfix ported components
 These components have removed the SCSS styling, but has not added it as JSS.
 
 Table, Tbody, Thead, Tfoot, Tr, Th, Td,
-Button, GraphTooltip, HorizontalNav, Legend, Pane,
+GraphTooltip, HorizontalNav, Legend, Pane,
 Dropdown, Alert, NavBar, Tooltip, Search,
 Spinner, Widget, RangeSelector, RatioBar, Select